Config notification ops to throw LookupError when using unknown id to lookup object

+++ src/backend/ops/notification.py
@@ -32,8 +32,10 @@ def query_channels(user_id=None, pending=None, skip=0, limit=None):
 
 # get channel for channel id
 # returns channel as a dict
+# throws LookupError if no channel with channel_id is found
 def get_channel(channel_id):
     channel = Channel.query.get(channel_id)
+    if channel is None: raise LookupError
     # map fields to dict
     return map_dict(channel, channel_mapping)
 
@@ -50,20 +52,26 @@ def create_channel(kind, user_id):
 # update channel with channel_id
 # kind - kind of channel (task/event/notice)
 # user_id - id of user using this channel
+# throws LookupError if no channel with channel_id is found
 def update_channel(channel_id, kind=None, user_id=None):
     channel = Channel.query.get(channel_id)
+    if channel is None: raise LookupError
+    # update channel fields
     if not kind is None: channel.kind = kind
     if not user_id is None: channel.user_id = user_id
     db.session.commit()
 
 # delete the channel with the given channel id
-# als cascade deletes any dependent objects
+# also cascade deletes any dependent objects
+# throws LookupError if no channel with channel_id is found
 def delete_channel(channel_id):
+    channel = Channel.query.get(channel_id)
+    if channel is None: raise LookupError
+
     # cascade delete notifications
     notify_ids = query_notify(channel_id=channel_id)
     for notify_id in notify_ids: delete_notify(notify_id)
 
-    channel = Channel.query.get(channel_id)
     db.session.delete(channel)
     db.session.commit()
 
@@ -89,8 +97,10 @@ def query_notify(pending=None, channel_id=None, skip=0, limit=None):
     return apply_bound(notify_id, skip, limit)
 
 # get notification for notify_id
+# throws LookupError if no notify with notify_id is found
 def get_notify(notify_id):
     notify = Notification.query.get(notify_id)
+    if notify is None: raise LookupError
     # map fields to dict
     return map_dict(notify, notify_mapping)
 
@@ -112,9 +122,12 @@ def create_notify(title, firing_time, channel_id, description=""):
 # firing_time - firing datetime of the notification
 # channel_id - id of the channel to send the notification
 # returns the id of the new notification
+# throws LookupError if no notify with notify_id is found
 def update_notify(notify_id, title=None, firing_time=None, channel_id=None,
                   description=None):
     notify = Notification.query.get(notify_id)
+    if notify is None: raise LookupError
+    # update notification fields
     if not title is None: notify.title = title
     if not firing_time is None: notify.firing_time = firing_time
     if not channel_id is None: notify.channel_id = channel_id
@@ -122,8 +135,9 @@ def update_notify(notify_id, title=None, firing_time=None, channel_id=None,
     db.session.commit()
 
 # delete notification with the given notify_id
+# throws LookupError if no notify with notify_id is found
 def delete_notify(notify_id):
     notify = Notification.query.get(notify_id)
+    if notify is None: raise LookupError
     db.session.delete(notify)
     db.session.commit()
-
